requirements.txt
chainlit
python-dotenv
requests
httpx
click
mcp[cli]
openai